src/kafka_init/requirements.txt
confluent-kafka==2.5.3
pyyaml==6.0.2
drgn==0.1.5